+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 Full changelog for the costs to expect REST API.
 
+## 2019-10-27 - v1.06.0
+
+* Updated database, a category is now a child of a resource type, not global.
+* Updated categories collection and category, shows the resource type that category is assigned to.
+* POST/resource_types/.../item/[item_id]/category updated to look at resource type.
+* POST/categories requires the resource_type_id to be set.
+* Request log and Request error log now show created times.
+* Minor updates to models.
+
 ## 2019-10-22 - v1.05.0
 
 * Added the ability to POST a request error to the API.